feat: extract shared Shipwright toolbar into @shipwright/lib
Both the admin UI and metrics dashboard now render the same navbar
(Agents, Provision, Metrics) from a single renderShipwrightToolbar()
in the new @shipwright/lib workspace. Admin and metrics call it with
their own logoutAction; metricsUrl defaults to /sw/dashboard.
