+++ admin/src/admin-ui-styles.ts
@@ -1,97 +1,13 @@
-/**
- * agent/src/admin-ui-styles.ts
- * Base CSS for the admin UI pages.
- * Extends the shared toolbar styles from metrics/src/lib/web/toolbar.ts.
- */
+import {
+  escapeHtml,
+  renderShipwrightToolbar,
+  baseStyles as toolbarBaseStyles,
+} from "@shipwright/lib/web/toolbar.ts";
+
+export { escapeHtml };
 
 export function baseStyles(): string {
-  return `
-    *, *::before, *::after { box-sizing: border-box; margin: 0; padding: 0; }
-    html { font-size: 16px; -webkit-font-smoothing: antialiased; }
-    body {
-      font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif;
-      background: #f7f7fb;
-      color: #1a1a2e;
-      min-height: 100vh;
-    }
-
-    /* ─── Toolbar ───────────────────────────────────────── */
-    .vos-toolbar {
-      background: #fff;
-      border-bottom: 1px solid #e8e8ee;
-      height: 52px;
-      display: flex;
-      align-items: center;
-      padding: 0 24px;
-      gap: 24px;
-      position: sticky;
-      top: 0;
-      z-index: 100;
-    }
-    .vos-wordmark {
-      font-size: 15px;
-      font-weight: 700;
-      color: #1a1a2e;
-      text-decoration: none;
-      letter-spacing: -0.3px;
-      flex-shrink: 0;
-    }
-    .vos-nav {
-      display: flex;
-      align-items: center;
-      gap: 4px;
-      flex: 1;
-    }
-    .vos-nav-link {
-      font-size: 13px;
-      font-weight: 500;
-      color: #6b7280;
-      text-decoration: none;
-      padding: 6px 10px;
-      border-radius: 6px;
-    }
-    .vos-nav-link:hover,
-    .vos-nav-link.active {
-      color: #1a1a2e;
-      background: #f3f4f6;
-    }
-    .vos-user {
-      display: flex;
-      align-items: center;
-      gap: 12px;
-      flex-shrink: 0;
-      margin-left: auto;
-    }
-    .vos-username {
-      font-size: 13px;
-      color: #6b7280;
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      max-width: 160px;
-    }
-    .vos-signout-btn {
-      font-size: 13px;
-      font-weight: 500;
-      color: #6b7280;
-      background: none;
-      border: 1px solid #e8e8ee;
-      border-radius: 6px;
-      padding: 5px 12px;
-      cursor: pointer;
-      font-family: inherit;
-    }
-    .vos-signout-btn:hover {
-      background: #f3f4f6;
-    }
-
-    /* ─── Page body ─────────────────────────────────────── */
-    .vos-page {
-      max-width: 960px;
-      margin: 0 auto;
-      padding: 32px 24px 64px;
-    }
-
+  return `${toolbarBaseStyles()}
     /* ─── Page header ───────────────────────────────────── */
     .page-header {
       display: flex;
@@ -359,30 +275,11 @@ export function baseStyles(): string {
   `;
 }
 
-export function escapeHtml(str: string): string {
-  return str
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;");
-}
-
 export function renderAdminToolbar(userName: string, activePath = ""): string {
-  const active = (prefix: string) =>
-    activePath.startsWith(prefix) ? " active" : "";
-  const metricsHref = process.env.METRICS_DASHBOARD_URL ?? "/sw/dashboard";
-  return `<nav class="vos-toolbar" aria-label="Site navigation">
-    <a href="/admin/agents" class="vos-wordmark">Shipwright Admin</a>
-    <div class="vos-nav">
-      <a href="/admin/agents" class="vos-nav-link${active("/admin/agents")}">Agents</a>
-      <a href="/admin/provision" class="vos-nav-link${active("/admin/provision")}">Provision</a>
-      <a href="${metricsHref}" class="vos-nav-link">Metrics</a>
-    </div>
-    <div class="vos-user">
-      <span class="vos-username">${escapeHtml(userName)}</span>
-      <form method="POST" action="/admin/logout" style="margin:0">
-        <button type="submit" class="vos-signout-btn">Sign out</button>
-      </form>
-    </div>
-  </nav>`;
+  return renderShipwrightToolbar({
+    userName,
+    activePath,
+    logoutAction: "/admin/logout",
+    metricsUrl: process.env.METRICS_DASHBOARD_URL ?? "/sw/dashboard",
+  });
 }
