+++ lib/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@shipwright/lib",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module"
+}
