+++ package.json
@@ -2,7 +2,7 @@
   "name": "shipwright-monorepo",
   "version": "0.1.0",
   "private": true,
-  "workspaces": ["plugins/shipwright", "metrics", "agent", "admin"],
+  "workspaces": ["lib", "plugins/shipwright", "metrics", "agent", "admin"],
   "scripts": {
     "test": "bun test",
     "lint": "bunx biome lint .",
